Adding 2 more enemies and adding capability to restrict enemies to start at certain waves. Stubbing out a music controller. Balancing towers.

+++ Assets/Scripts/TowerController.cs
@@ -13,7 +13,7 @@ public class TowerController : MonoBehaviour
     public int PlaceCost;
     public int UpgradeCost;
 
-    public int Range;
+    public float Range;
     public int Damage;
     public float AttackSpeed;
 
@@ -173,7 +173,7 @@ public class TowerController : MonoBehaviour
         Invoke("CheckAttack", GetAttackSpeed());
     }
 
-    public int GetRange()
+    public float GetRange()
     {
         return Range + ((level - 1) * UpgradeRange);
     }
